Changed problem->evaluations and problem->best_observed_evaluation from long to size_t (to be consistent).

+++ code-experiments/src/coco_generics.c
@@ -19,7 +19,7 @@ void coco_evaluate_function(coco_problem_t *self, const double *x, double *y) {
 #endif
 }
 
-long coco_problem_get_evaluations(coco_problem_t *self) {
+size_t coco_problem_get_evaluations(coco_problem_t *self) {
   assert(self != NULL);
   return self->evaluations;
 }
